--- monai/networks/blocks/attention.py
@@ -1,327 +1,0 @@
-# Copyright (c) MONAI Consortium
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#     http://www.apache.org/licenses/LICENSE-2.0
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import annotations
-
-from typing import Optional, Tuple
-
-import torch
-import torch.nn as nn
-
-from monai.networks.layers.utils import get_rel_pos_embedding_layer
-from monai.utils import optional_import
-
-Rearrange, _ = optional_import("einops.layers.torch", name="Rearrange")
-
-
-class SABlock(nn.Module):
-    """
-    A self-attention block, based on: "Dosovitskiy et al.,
-    An Image is Worth 16x16 Words: Transformers for Image Recognition at Scale <https://arxiv.org/abs/2010.11929>"
-    One can setup relative positional embedding as described in <https://arxiv.org/abs/2112.01526>
-    """
-
-    def __init__(
-        self,
-        hidden_size: int,
-        num_heads: int,
-        dropout_rate: float = 0.0,
-        hidden_input_size: int | None = None,
-        num_head_channels: int | None = None,
-        qkv_bias: bool = False,
-        save_attn: bool = False,
-        causal: bool = False,
-        sequence_length: int | None = None,
-        rel_pos_embedding: Optional[str] = None,
-        input_size: Optional[Tuple] = None,
-    ) -> None:
-        """
-        Args:
-            hidden_size (int): dimension of hidden layer.
-            num_heads (int): number of attention heads.
-            dropout_rate (float, optional): fraction of the input units to drop. Defaults to 0.0.
-            qkv_bias (bool, optional): bias term for the qkv linear layer. Defaults to False.
-            causal: whether to use causal attention.
-            sequence_length: if causal is True, it is necessary to specify the sequence length.
-            rel_pos_embedding (str, optional): Add relative positional embeddings to the attention map.
-                For now only "decomposed" is supported (see https://arxiv.org/abs/2112.01526). 2D and 3D are supported.
-            input_size (tuple(spatial_dim), optional): Input resolution for calculating the relative
-                positional parameter size.
-            save_attn (bool, optional): to make accessible the attention matrix. Defaults to False.
-
-        """
-
-        super().__init__()
-
-        if not (0 <= dropout_rate <= 1):
-            raise ValueError("dropout_rate should be between 0 and 1.")
-
-        if num_head_channels:
-            inner_size = num_heads * num_head_channels
-            self.head_dim = num_head_channels
-        else:
-            if hidden_size % num_heads != 0:
-                raise ValueError("hidden size should be divisible by num_heads.")
-            inner_size = hidden_size
-            self.head_dim = hidden_size // num_heads
-
-        if causal and sequence_length is None:
-            raise ValueError("sequence_length is necessary for causal attention.")
-
-        self.num_heads = num_heads
-        self.hidden_input_size = hidden_input_size if hidden_input_size else hidden_size
-        self.out_proj = nn.Linear(inner_size, self.hidden_input_size)
-        self.qkv = nn.Linear(self.hidden_input_size, inner_size * 3, bias=qkv_bias)
-        self.input_rearrange = Rearrange("b h (qkv l d) -> qkv b l h d", qkv=3, l=num_heads)
-        self.out_rearrange = Rearrange("b h l d -> b l (h d)")
-        self.drop_output = nn.Dropout(dropout_rate)
-        self.drop_weights = nn.Dropout(dropout_rate)
-        self.scale = self.head_dim**-0.5
-        self.save_attn = save_attn
-        self.causal = causal
-        self.sequence_length = sequence_length
-
-        if causal and sequence_length is not None:
-            # causal mask to ensure that attention is only applied to the left in the input sequence
-            self.register_buffer(
-                "causal_mask",
-                torch.tril(torch.ones(sequence_length, sequence_length)).view(1, 1, sequence_length, sequence_length),
-            )
-            self.causal_mask: torch.Tensor
-
-        self.att_mat = torch.Tensor()
-        self.rel_positional_embedding = (
-            get_rel_pos_embedding_layer(rel_pos_embedding, input_size, self.head_dim, self.num_heads)
-            if rel_pos_embedding is not None
-            else None
-        )
-        self.input_size = input_size
-
-    def forward(self, x: torch.Tensor):
-        """
-        Args:
-            x (torch.Tensor): input tensor. B x (s_dim_1 * ... * s_dim_n) x C
-
-        Return:
-            torch.Tensor: B x (s_dim_1 * ... * s_dim_n) x C
-        """
-        output = self.input_rearrange(self.qkv(x))
-        q, k, v = output[0], output[1], output[2]
-        att_mat = torch.einsum("blxd,blyd->blxy", q, k) * self.scale
-
-        # apply relative positional embedding if defined
-        att_mat = self.rel_positional_embedding(x, att_mat, q) if self.rel_positional_embedding is not None else att_mat
-
-        if self.causal:
-            att_mat = att_mat.masked_fill(self.causal_mask[:, :, : x.shape[1], : x.shape[1]] == 0, float("-inf"))
-
-        att_mat = att_mat.softmax(dim=-1)
-
-        if self.save_attn:
-            # no gradients and new tensor;
-            # https://pytorch.org/docs/stable/generated/torch.Tensor.detach.html
-            self.att_mat = att_mat.detach()
-
-        att_mat = self.drop_weights(att_mat)
-        x = torch.einsum("bhxy,bhyd->bhxd", att_mat, v)
-        x = self.out_rearrange(x)
-        x = self.out_proj(x)
-        x = self.drop_output(x)
-        return x
-
-
-class CrossAttentionBlock(nn.Module):
-    """
-    A cross-attention block, based on: "Dosovitskiy et al.,
-    An Image is Worth 16x16 Words: Transformers for Image Recognition at Scale <https://arxiv.org/abs/2010.11929>"
-    One can setup relative positional embedding as described in <https://arxiv.org/abs/2112.01526>
-    """
-
-    def __init__(
-        self,
-        hidden_size: int,
-        num_heads: int,
-        hidden_input_size: int | None = None,
-        context_input_size: int | None = None,
-        num_head_channels: int | None = None,
-        dropout_rate: float = 0.0,
-        qkv_bias: bool = False,
-        save_attn: bool = False,
-        causal: bool = False,
-        sequence_length: int | None = None,
-        rel_pos_embedding: Optional[str] = None,
-        input_size: Optional[Tuple] = None,
-    ) -> None:
-        """
-        Args:
-            hidden_size (int): dimension of hidden layer.
-            num_heads (int): number of attention heads.
-            dropout_rate (float, optional): fraction of the input units to drop. Defaults to 0.0.
-            qkv_bias (bool, optional): bias term for the qkv linear layer. Defaults to False.
-            causal: whether to use causal attention.
-            sequence_length: if causal is True, it is necessary to specify the sequence length.
-            rel_pos_embedding (str, optional): Add relative positional embeddings to the attention map.
-                For now only "decomposed" is supported (see https://arxiv.org/abs/2112.01526). 2D and 3D are supported.
-            input_size (tuple(spatial_dim), optional): Input resolution for calculating the relative
-                positional parameter size.
-            save_attn (bool, optional): to make accessible the attention matrix. Defaults to False.
-
-        """
-
-        super().__init__()
-
-        if not (0 <= dropout_rate <= 1):
-            raise ValueError("dropout_rate should be between 0 and 1.")
-
-        if num_head_channels:
-            inner_size = num_heads * num_head_channels
-            self.head_dim = num_head_channels
-        else:
-            if hidden_size % num_heads != 0:
-                raise ValueError("hidden size should be divisible by num_heads.")
-            inner_size = hidden_size
-            self.head_dim = hidden_size // num_heads
-
-        if causal and sequence_length is None:
-            raise ValueError("sequence_length is necessary for causal attention.")
-
-        self.num_heads = num_heads
-        self.hidden_input_size = hidden_input_size if hidden_input_size else hidden_size
-        self.context_input_size = context_input_size if context_input_size else hidden_size
-        self.out_proj = nn.Linear(inner_size, self.hidden_input_size)
-        # key, query, value projections
-        self.to_q = nn.Linear(self.hidden_input_size, inner_size, bias=qkv_bias)
-        self.to_k = nn.Linear(self.context_input_size, inner_size, bias=qkv_bias)
-        self.to_v = nn.Linear(self.context_input_size, inner_size, bias=qkv_bias)
-        self.input_rearrange = Rearrange("b h (l d) -> b l h d", l=num_heads)
-
-        self.out_rearrange = Rearrange("b h l d -> b l (h d)")
-        self.drop_output = nn.Dropout(dropout_rate)
-        self.drop_weights = nn.Dropout(dropout_rate)
-
-        self.scale = self.head_dim**-0.5
-        self.save_attn = save_attn
-        self.causal = causal
-        self.sequence_length = sequence_length
-
-        if causal and sequence_length is not None:
-            # causal mask to ensure that attention is only applied to the left in the input sequence
-            self.register_buffer(
-                "causal_mask",
-                torch.tril(torch.ones(sequence_length, sequence_length)).view(1, 1, sequence_length, sequence_length),
-            )
-            self.causal_mask: torch.Tensor
-
-        self.att_mat = torch.Tensor()
-        self.rel_positional_embedding = (
-            get_rel_pos_embedding_layer(rel_pos_embedding, input_size, self.head_dim, self.num_heads)
-            if rel_pos_embedding is not None
-            else None
-        )
-        self.input_size = input_size
-
-    def forward(self, x: torch.Tensor, context: torch.Tensor | None = None):
-        """
-        Args:
-            x (torch.Tensor): input tensor. B x (s_dim_1 * ... * s_dim_n) x C
-            context (torch.Tensor, optional): context tensor. B x (s_dim_1 * ... * s_dim_n) x C
-
-        Return:
-            torch.Tensor: B x (s_dim_1 * ... * s_dim_n) x C
-        """
-        # calculate query, key, values for all heads in batch and move head forward to be the batch dim
-        b, t, c = x.size()  # batch size, sequence length, embedding dimensionality (hidden_size)
-
-        q = self.to_q(x)
-        kv = context if context is not None else x
-        _, kv_t, _ = kv.size()
-        k = self.to_k(kv)
-        v = self.to_v(kv)
-        q = q.view(b, t, self.num_heads, c // self.num_heads).transpose(1, 2)  # (b, nh, t,  hs)
-        k = k.view(b, kv_t, self.num_heads, c // self.num_heads).transpose(1, 2)  # (b, nh, kv_t, hs)
-        v = v.view(b, kv_t, self.num_heads, c // self.num_heads).transpose(1, 2)  # (b, nh, kv_t, hs)
-        att_mat = torch.einsum("blxd,blyd->blxy", q, k) * self.scale
-
-        # apply relative positional embedding if defined
-        att_mat = self.rel_positional_embedding(x, att_mat, q) if self.rel_positional_embedding is not None else att_mat
-
-        if self.causal:
-            att_mat = att_mat.masked_fill(self.causal_mask[:, :, :t, :kv_t] == 0, float("-inf"))
-
-        att_mat = att_mat.softmax(dim=-1)
-
-        if self.save_attn:
-            # no gradients and new tensor;
-            # https://pytorch.org/docs/stable/generated/torch.Tensor.detach.html
-            self.att_mat = att_mat.detach()
-
-        att_mat = self.drop_weights(att_mat)
-        x = torch.einsum("bhxy,bhyd->bhxd", att_mat, v)
-        x = self.out_rearrange(x)
-        x = self.out_proj(x)
-        x = self.drop_output(x)
-        return x
-
-
-class SpatialAttentionBlock(nn.Module):
-    """Perform spatial self-attention on the input tensor.
-
-    The input tensor is reshaped to B x (x_dim * y_dim [ * z_dim]) x C, where C is the number of channels, and then
-    self-attention is performed on the reshaped tensor. The output tensor is reshaped back to the original shape.
-
-    Args:
-        spatial_dims: number of spatial dimensions, could be 1, 2, or 3.
-        num_channels: number of input channels. Must be divisible by num_head_channels.
-        num_head_channels: number of channels per head.
-    """
-
-    def __init__(
-        self,
-        spatial_dims: int,
-        num_channels: int,
-        num_head_channels: int | None = None,
-        norm_num_groups: int = 32,
-        norm_eps: float = 1e-6,
-    ) -> None:
-        super().__init__()
-
-        self.spatial_dims = spatial_dims
-        self.norm = nn.GroupNorm(num_groups=norm_num_groups, num_channels=num_channels, eps=norm_eps, affine=True)
-        # check num_head_channels is divisible by num_channels
-        if num_head_channels is not None and num_channels % num_head_channels != 0:
-            raise ValueError("num_channels must be divisible by num_head_channels")
-        num_heads = num_channels // num_head_channels if num_head_channels is not None else 1
-
-        self.attn = SABlock(hidden_size=num_channels, num_heads=num_heads, qkv_bias=True)
-
-    def forward(self, x: torch.Tensor):
-        residual = x
-
-        if self.spatial_dims == 1:
-            h = x.shape[2]
-            rearrange_input = Rearrange("b c h -> b h c")
-            rearrange_output = Rearrange("b h c -> b c h", h=h)
-        if self.spatial_dims == 2:
-            h, w = x.shape[2], x.shape[3]
-            rearrange_input = Rearrange("b c h w -> b (h w) c")
-            rearrange_output = Rearrange("b (h w) c -> b c h w", h=h, w=w)
-        if self.spatial_dims == 3:
-            h, w, d = x.shape[2], x.shape[3], x.shape[4]
-            rearrange_input = Rearrange("b c h w d -> b (h w d) c")
-            rearrange_output = Rearrange("b (h w d) c -> b c h w d", h=h, w=w, d=d)
-
-        x = self.norm(x)
-        x = rearrange_input(x)  # B x (x_dim * y_dim [ * z_dim]) x C
-
-        x = self.attn(x)
-        x = rearrange_output(x)  # B x  x C x x_dim * y_dim * [z_dim]
-        x = x + residual
-        return x
